Fixed find and find_index
